refactor(query): Split runner/query.rs into submodules
